(build) fix warnings from logging type mismatch
uint32_ts passed as size_ts and vice versa.

FossilOrigin-Name: 0947d960c291134a34e8a4f05d8cc8bc34242a34b2f427d37b55ee7469a4a7a3

+++ src/a12/a12_encode.c
@@ -436,7 +436,8 @@ static void compress_tz(struct a12_state* S, uint8_t ch,
 
 		a12int_trace(A12_TRACE_VDETAIL,
 			"kind=status:codec=dzstd:b_in=%zu:b_out=%zu:ratio=%.2f",
-			compress_in_sz, out_sz, (float)(compress_in_sz+1.0) / (float)(out_sz+1.0)
+			(size_t)compress_in_sz,
+			(size_t) out_sz, (float)(compress_in_sz+1.0) / (float)(out_sz+1.0)
 		);
 	}
 	else
@@ -456,7 +457,7 @@ static void compress_tz(struct a12_state* S, uint8_t ch,
 
 	a12int_trace(A12_TRACE_VDETAIL,
 		"kind=status:codec=tpack:b_in=%zu:b_out=%zu:zstd=%d",
-		compress_in_sz, compress_in_sz, out_sz, (int) zstd
+		(size_t) compress_in_sz, (size_t) compress_in_sz, (size_t) out_sz, (int)zstd
 	);
 
 	a12int_append_out(S,
